int_float
Former-commit-id: 2a184be2092595c58cce21cae3ba883804b9858d

+++ cpp/2016_expression/literal.h
@@ -7,7 +7,7 @@
 class Literal : public Expression
 {
 private:
-    int val_;
+    float val_;
 public:
     Literal(float val);
     float eval();
